requirement.txt
asgiref==3.8.1
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
Django==5.2.3
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.29.0
google-api-python-client==2.111.0
google-auth==2.47.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
oauthlib==3.3.1
proto-plus==1.27.0
protobuf==6.33.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.3.1
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3